Update kendra_chat_bedrock_claudev2.py
The order of the history had to be changed because the Bedrock API expects to see alternating Human:/Assistant: lines for Claude.  Without this change, the first query would work, but subsequent queries would generate an error from the Bedrock API:  

ValueError: Error: Prompt must alternate between '

Human:' and '

Assistant:'.